requirements.txt
requests
binance-connector
pandas
numpy
sqlalchemy
psycopg2-binary
boto3
python-dotenv